Re-activated fish-eye in heat-map
git-svn-id: svn://localhost:8080/caleydo_main/trunk@4554 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.view.heatmap/src/org/caleydo/view/heatmap/heatmap/renderer/spacing/ASpacingCalculator.java
@@ -7,7 +7,7 @@ public abstract class ASpacingCalculator {
 	protected GLHeatMap heatMap;
 	protected float y;
 	protected float recordElements;
-	protected float minSelectedFieldHeight;
+	protected int minSelectedFieldHeight;
 
 	public ASpacingCalculator(GLHeatMap heatMap, float y, float recordElements) {
 		this.heatMap = heatMap;
@@ -19,7 +19,7 @@ public abstract class ASpacingCalculator {
 
 	public abstract float getFieldHeight(int recordID);
 
-	public void setMinSelectedFieldHeight(float minSelectedFieldHeight) {
+	public void setMinSelectedFieldHeight(int minSelectedFieldHeight) {
 		this.minSelectedFieldHeight = minSelectedFieldHeight;
 	}
